Rename project lifecycle stage 'Adopted' to 'Graduated'
This has been a discussion topic for some time, so it likely makes sense to codify this.

Two reasons for this change:

- The term 'Adopted' ends up being a misnomer, making it seem that projects like OTIO or OIIO that have been used for eons by the industry aren't/shouldn't be adopted. The TAC lifecycle views maturity in terms of open source best practices and infrastructure ( which can be a leading indicator of adoption, but not a direct correlation )
'Graduated' is a more common term used in other project lifecycles, so this helps align expectations across ecosystems.

Please review/discuss. I've also changed the 'Graduated' stage description to clarify it better.

Signed-off-by: John Mertic <jmertic@linuxfoundation.org>